Use <div>, <a>, <form>, <nav>
The <div> element is the generic container for flow content. It has no effect on the content or layout until styled in some way using CSS.
The <a> element (or anchor element), with its href attribute, creates a hyperlink to web pages, files, email addresses, locations in the same page, or anything else a URL can address.
The <form> element represents a document section containing interactive controls for submitting information.
The <nav> element represents a section of a page whose purpose is to provide navigation links, either within the current document or to other documents.